Major changes to Bridge Component and Player movements
Completely redesigned getTop, GetMid, GetBot methods
Added custom offsets for terrainfactory grids and screen grids
Added Rescale between grid layouts

+++ source/core/src/main/com/deco2800/game/areas/RainbowBridge.java
@@ -151,7 +151,7 @@ public class RainbowBridge extends GameArea {
 
         for (int i = 0; i < lanes.size(); i++) {
             int d = 0;
-            int y_coordinate = lanes.get(i).getMid() ;
+            int y_coordinate = (int) lanes.get(i).getMid() ;
             int x_random = ThreadLocalRandom.current().nextInt(0, 30 + 1);
             GridPoint2 randomPosInLane = new GridPoint2(x_random, y_coordinate);
             // Entity RunesGate = ObstacleFactory.createRunesGate();
@@ -184,7 +184,7 @@ public class RainbowBridge extends GameArea {
 
         for (int i = 0; i < lanes.size(); i++) {
             int d = 0;
-            int y_coordinate = lanes.get(i).getMid();
+            int y_coordinate = (int) lanes.get(i).getMid();
             int x_random = ThreadLocalRandom.current().nextInt(5, 30 + 1);  // min x=5, max x=30
             GridPoint2 randomPosInLane = new GridPoint2(x_random, y_coordinate);
         
@@ -205,7 +205,7 @@ public class RainbowBridge extends GameArea {
         List<Lane> lanes = terrain.getRainbowBridge().getLanes();
         int i = 0;
         for (Lane lane : lanes) {
-            int y_coordinate = lane.getMid();
+            int y_coordinate = (int) lane.getMid();
             int x_random = ThreadLocalRandom.current().nextInt(5, 28 + 1);  // min x=5, max x=28
             GridPoint2 randomPosInLane = new GridPoint2(x_random, y_coordinate);
         
@@ -232,7 +232,7 @@ public class RainbowBridge extends GameArea {
 
         for (int i = 0; i < lanes.size(); i++) {
             int d = 0;
-            int y_coordinate = lanes.get(i).getMid();
+            int y_coordinate = (int) lanes.get(i).getMid();
             int x_random = ThreadLocalRandom.current().nextInt(5, 28 + 1);  // min x=5, max x=28
             GridPoint2 randomPosInLane = new GridPoint2(x_random, y_coordinate);
